Add support for analyzing GopherJS artifact size.
When --analyze_size flag is passed to GopherJS, an additional
${packagename}.js.svg file is creates with a TreeMap diagram that
visualizes which packages contributed to the artifact size and how much.

The data is obtained by counting bytes written for each section of the
file including GopherJS's own prelude and works accurately both in
minified and default modes.

Most of diagram-related logic is isolated to the build/analysis package
with minimal changes to the rest of the code base.